A new Utility Class to ensure correct unit conversion.
This class has static members of scientific constants and converts from
qunatities from one unit system to another. Currently only a few simple
conversion between NU and SI are implented, feel free to add more.
The point is to ensure correct unit conversion as it is easy to mix up
small details which cause very wrong results.
All function take 3 arguments, a value that is to be converted, a
metricprefix for the input and the desired one for the output. The
arguments have default values that should correspond to the way
quantities are stored in root files (e.g. Energy conversion assumes GeV
and Joule).